A simple Hello World example that binds a templated view to a model.
In main.js a Person is defined, a Person just extends Ember.Object with a
property called name.

A View called HelloView is defined, and hooked up to the template named
'hello-view'. The template is declared in a script block in the markup and
named 'hello-view'.

An instance of the model is created with the name 'Bob' and this model is
passed to new instance of HelloView and appended to the document body.